@@ -11,7 +11,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 20140724192131) do
+ActiveRecord::Schema.define(version: 20140726154911) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
@@ -37,6 +37,8 @@ ActiveRecord::Schema.define(version: 20140724192131) do
     t.datetime "updated_at"
     t.integer  "median_close_time"
     t.integer  "issues_count"
+    t.text     "pr_distribution"
+    t.text     "issues_distribution"
   end
 
 end
