@@ -2,7 +2,6 @@ require 'rubygems'
 ENV["RAILS_ENV"] ||= 'test'
 require File.expand_path("../../config/environment", __FILE__)
 require 'rspec/rails'
-require 'rspec/autorun'
 require 'capybara/rspec'
 require 'database_cleaner'
 Dir[Rails.root.join("spec/support/**/*.rb")].each {|f| require f}
@@ -42,4 +41,5 @@ RSpec.configure do |config|
 
   config.infer_base_class_for_anonymous_controllers = false
   config.order = "random"
+  # config.raise_errors_for_deprecations!
 end
